Renamed deploy verifier for custom smart wallets. Pull up runtimeCodeHash() function to IWalletFactory to remove unneded relay verifier for custom smart wallets. Extracted EnvelopingUtils from Utils to its own file and renamed to Enveloping. Updated documentation.
