+++ contracts/interfaces/IWalletFactory.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier:MIT
 pragma solidity ^0.6.12;
 pragma experimental ABIEncoderV2;
 
@@ -7,6 +7,8 @@ import "./IForwarder.sol";
 interface IWalletFactory {
 
     function nonce (address from) external view returns(uint256);
+
+    function runtimeCodeHash() external view returns (bytes32);
 
     function relayedUserSmartWalletCreation(
         IForwarder.DeployRequest memory req,
