+++ contracts/test/ERC20Mod.sol
@@ -1,5 +1,4 @@
 // SPDX-License-Identifier: MIT
-
 pragma solidity ^0.6.12;
 
 import "@openzeppelin/contracts/GSN/Context.sol";
