+++ contracts/test/TestVersions.sol
@@ -2,12 +2,6 @@
 pragma solidity ^0.6.12;
 
 contract TestVersions {
-
-    // stub for ContractInteractor "init" to succeed
-    function stakeManager() external view returns (address) {
-        return address(this);
-    }
-
     function versionHub() external pure returns (string memory) {
         return "3.0.0";
     }
