+++ contracts/utils/MinLibBytes.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier:MIT
 // minimal bytes manipulation required by Enveloping
 // a minimal subset from 0x/LibBytes
 /* solhint-disable no-inline-assembly */
