+++ src/relayclient/KnownRelaysManager.ts
@@ -68,7 +68,7 @@ export class KnownRelaysManager {
     }
     const topics = addresses2topics(Array.from(relayManagers))
     const relayServerRegisteredEvents = await this.contractInteractor.getPastEventsForHub(topics, { fromBlock: 1 }, [RelayServerRegistered])
-    const relayManagerExitEvents = await this.contractInteractor.getPastEventsForStakeManager([StakeUnlocked, StakePenalized], topics, { fromBlock: 1 })
+    const relayManagerExitEvents = await this.contractInteractor.getPastEventsForStakeManagement([StakeUnlocked, StakePenalized], topics, { fromBlock: 1 })
 
     log.info(`== fetchRelaysAdded: found ${relayServerRegisteredEvents.length} unique RelayAdded events`)
